backend/requirements.txt
stripe==6.1.0
# aniso8601==9.0.1
# click==8.0.3
Flask==2.2.1
Flask-HTTPAuth==4.5.0
Flask-RESTful==0.3.9
# itsdangerous==2.0.1
# Jinja2==3.1.2
# MarkupSafe==2.1.1
# pytz==2021.3
# six==1.16.0
# Werkzeug==2.2.0
Flask-Cors==3.0.10
Flask-SQLAlchemy==3.0.3
python-dotenv==1.0.0